chore: Move confirmRemovingServices to its own file
