+++ Wire-iOS/Sources/UserInterface/Conversation Options/UIAlertController+ConfirmRemovingGuests.swift
@@ -39,14 +39,6 @@ extension UIAlertController {
         )
     }
 
-    static func confirmRemovingServices(_ completion: @escaping (Bool) -> Void) -> UIAlertController {
-        return confirmController(
-            title: "services_room.remove_services.message".localized,
-            confirmTitle: "services_room.remove_services.action".localized,
-            completion: completion
-        )
-    }
-
     static func confirmRevokingLink(_ completion: @escaping (Bool) -> Void) -> UIAlertController {
         return confirmController(
             title: "guest_room.revoke_link.message".localized,
